+++ src-java/flow-topology/flow-messaging/src/main/java/org/openkilda/messaging/command/flow/FlowRequest.java
@@ -58,7 +58,7 @@ public class FlowRequest extends CommandData {
     boolean allocateProtectedPath;
     String description;
     int transitEncapsulationId;
-    Integer maxLatency;
+    Long maxLatency;
     Integer priority;
     boolean pinned;
     String diverseFlowId;
